Extract Class From PlayCommand (Add MusicLoader)
