+++ src/main/java/com/discord/bot/commands/musiccommands/PlayCommand.java
@@ -1,7 +1,8 @@
 package com.discord.bot.commands.musiccommands;
 
 import com.discord.bot.audioplayer.GuildMusicManager;
-import com.discord.bot.dao.pojo.MusicPojo;
+import com.discord.bot.entity.pojo.MusicPojo;
+import com.discord.bot.loader.MusicLoader;
 import com.discord.bot.service.RestService;
 import com.discord.bot.service.TrackService;
 import com.discord.bot.service.audioplayer.PlayerManagerService;
@@ -17,6 +18,7 @@ import java.util.List;
 public class PlayCommand extends MusicPlayerCommand {
     private TrackService trackService;
     private RestService restService;
+    private MusicLoader musicLoader;
 
     public PlayCommand(RestService restService, PlayerManagerService playerManagerService, TrackService trackService, MusicCommandUtils utils) {
         this.restService = restService;
@@ -28,7 +30,7 @@ public class PlayCommand extends MusicPlayerCommand {
     @Override
     public void execute(SlashCommandInteractionEvent event) {
         String query = event.getOption("query").getAsString().trim();
-        List<MusicPojo> musicPojos = getYoutubeLink(query, event);
+        List<MusicPojo> musicPojos = musicLoader.getMusicPojos(query, event.getChannel());
         playMusic(event, musicPojos);
     }
 
@@ -68,34 +70,5 @@ public class PlayCommand extends MusicPlayerCommand {
         } else {
             event.replyEmbeds(new EmbedBuilder().setDescription("Please join to a voice channel.").build()).queue();
         }
-    }
-
-    private List<MusicPojo> getYoutubeLink(String query, SlashCommandInteractionEvent event) {
-        List<MusicPojo> musicPojos = new ArrayList<>();
-        MusicPojo musicPojo;
-        if (query.contains("https://www.youtube.com/watch?v=")) {
-            musicPojo = new MusicPojo(null, query);
-            musicPojos.add(musicPojo);
-        } else if (query.contains("https://open.spotify.com/")) {
-            musicPojos = spotifyToYoutube(query);
-        } else {
-            musicPojo = restService.getYoutubeLink(new MusicPojo(query, null));
-            if (musicPojo.getYoutubeUri().equals("403glaxierror")) {
-                apiLimitExceeded(event.getChannel());
-            } else {
-                musicPojos.add(musicPojo);
-            }
-        }
-        return musicPojos;
-    }
-
-    private List<MusicPojo> spotifyToYoutube(String spotifyUrl) {
-        List<MusicPojo> musicPojoNameList = restService.getSpotifyMusicName(spotifyUrl);
-        return musicPojoNameList;
-    }
-
-    private void apiLimitExceeded(MessageChannel channel) {
-        channel.sendMessageEmbeds(new EmbedBuilder().setDescription("Youtube quota has exceeded. " +
-                "Please use youtube links to play music for today.").build()).queue();
     }
 }
